Principio solid - Principio da responsabilidade unica
