+++ src/app/controllers/FormController.js
@@ -4,6 +4,7 @@ import StoreDependentService from '../services/StoreDependentService';
 import GetUserDependentsFormService from '../services/GetUserDependentsFormService';
 import GetUserFormService from '../services/GetUserFormService';
 import validateProviderPerson from '../../validations/ProviderPersonValidation';
+import validateDependent from '../../validations/DependentValidation';
 import AppError from '../../errors/AppError';
 
 class FormController {
@@ -19,29 +20,7 @@ class FormController {
   }
 
   async storeDependent(req, res) {
-    const schema = Yup.object().shape({
-      // user
-      name: Yup.string().required(),
-
-      // person
-      document_number: Yup.string().required(),
-      birth_date: Yup.date().required(),
-      nationality: Yup.string().required(),
-      birth_city: Yup.string().required(),
-      birth_state: Yup.string().required(),
-      sex: Yup.string().required(),
-      breed: Yup.string().required(),
-      mother_name: Yup.string().required(),
-      father_name: Yup.string().required(),
-
-      // phone
-      phone_number: Yup.string().required(),
-      phone_code: Yup.string().required(),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      throw new AppError('Está faltando algum valor', 400);
-    }
+    await validateDependent(req.body);
 
     const storeDependentService = new StoreDependentService();
     await storeDependentService.execute(req.body);
